feat(queues): simplify timer handling and remove garbage collection
Previously, queues were run by an interval timer that periodically
checked whether new items had been queue or whether the queue was idle
long enough to be deleted.  With this change, queues are run by a
one-shot timer that runs as long as there are entries to be sent.
Once the queue is empty, it is deleted, removing the need for garbage
collection.